♻️ Refactor: CommunityVPAdapter 위치 변경
Related to: #303
