+++ app/src/main/java/com/mongmong/namo/presentation/ui/community/CommunityVPAdapter.kt
@@ -1,4 +1,4 @@
-package com.mongmong.namo.presentation.ui.community.adapter
+package com.mongmong.namo.presentation.ui.community
 
 import androidx.fragment.app.Fragment
 import androidx.viewpager2.adapter.FragmentStateAdapter
